Fix connection issues and make the callback failure more verbose
Fixed the connection error. It turns out that I was sending a connection packet instead of a connection successfull packet. This caused me to fix the wait for callback method silently failing. Currently this has the byproduct of only dealing with the first packet recieved and not waiting the timeout however this should fix any confusion that was caused for future use